backend/requirements.txt
# Django & DRF
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.0

# Database
psycopg2-binary>=2.9

# PDF Extraction
pdfminer.six>=20231228
PyMuPDF>=1.24
pytesseract>=0.3.10
Pillow>=10.2

# NLP
spacy>=3.7
nltk>=3.8
transformers>=4.38
sentence-transformers>=2.5
torch>=2.2

# ML Models
scikit-learn>=1.4
xgboost>=2.0
autogluon.tabular>=1.1
shap>=0.45
lime>=0.2

# Feature Engineering
numpy>=1.26
pandas>=2.2
scipy>=1.12

# Async Tasks
celery>=5.3
redis>=5.0

# Utils
python-dotenv>=1.0
gunicorn>=21.2
